+++ crates/nostr-database/src/lib.rs
@@ -102,7 +102,7 @@ where
 /// Nostr Database
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-pub trait NostrDatabase: AsyncTraitDeps {
+pub trait NostrDatabase: fmt::Debug + Send + Sync {
     /// Error
     type Err: From<DatabaseError> + Into<DatabaseError>;
 
@@ -357,37 +357,3 @@ impl<T: NostrDatabase> NostrDatabase for EraseNostrDatabaseError<T> {
         self.0.wipe().await.map_err(Into::into)
     }
 }
-
-/// Alias for `Send` on non-wasm, empty trait (implemented by everything) on
-/// wasm.
-#[cfg(not(target_arch = "wasm32"))]
-pub trait SendOutsideWasm: Send {}
-#[cfg(not(target_arch = "wasm32"))]
-impl<T: Send> SendOutsideWasm for T {}
-
-/// Alias for `Send` on non-wasm, empty trait (implemented by everything) on
-/// wasm.
-#[cfg(target_arch = "wasm32")]
-pub trait SendOutsideWasm {}
-#[cfg(target_arch = "wasm32")]
-impl<T> SendOutsideWasm for T {}
-
-/// Alias for `Sync` on non-wasm, empty trait (implemented by everything) on
-/// wasm.
-#[cfg(not(target_arch = "wasm32"))]
-pub trait SyncOutsideWasm: Sync {}
-#[cfg(not(target_arch = "wasm32"))]
-impl<T: Sync> SyncOutsideWasm for T {}
-
-/// Alias for `Sync` on non-wasm, empty trait (implemented by everything) on
-/// wasm.
-#[cfg(target_arch = "wasm32")]
-pub trait SyncOutsideWasm {}
-#[cfg(target_arch = "wasm32")]
-impl<T> SyncOutsideWasm for T {}
-
-/// Super trait that is used for our store traits, this trait will differ if
-/// it's used on WASM. WASM targets will not require `Send` and `Sync` to have
-/// implemented, while other targets will.
-pub trait AsyncTraitDeps: fmt::Debug + SendOutsideWasm + SyncOutsideWasm {}
-impl<T: fmt::Debug + SendOutsideWasm + SyncOutsideWasm> AsyncTraitDeps for T {}